requirements.txt
Faker==11.3.0
pynacl==1.4.0